Refactor #1 - Separte ToDo feature from Page Component by Creating its own
Since Page is the component handling entire screen. Separate ToDo Feature as independent Component would make it part of the screen.
