+++ src/Page/Page.jsx
@@ -1,25 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-import Form from '../commons/Form';
-import List from '../components/List';
+import ToDo from '../components/ToDo';
 
 function Page() {
-  const [list, setList] = useState([]);
-
-  function appendTask(value) {
-    setList([...list, value]);
-  }
-
-  function detachTask(id) {
-    setList(list.filter((task) => task.id !== id));
-  }
-
   return (
-    <div>
-      <h1>To-do</h1>
-      <Form appendTask={appendTask} />
-      <List list={list} onClick={detachTask} />
-    </div>
+    <ToDo />
   );
 }
